move merge-proj code into Project.java
