+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/algebra/Project.java
@@ -21,7 +21,10 @@ import org.apache.doris.nereids.trees.expressions.Alias;
 import org.apache.doris.nereids.trees.expressions.Expression;
 import org.apache.doris.nereids.trees.expressions.NamedExpression;
 import org.apache.doris.nereids.trees.expressions.Slot;
+import org.apache.doris.nereids.trees.expressions.SlotReference;
+import org.apache.doris.nereids.trees.expressions.visitor.DefaultExpressionRewriter;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -55,4 +58,79 @@ public interface Project {
                         )
                 );
     }
+
+    /**
+     * combine upper level and bottom level projections
+     * 1. alias combination, for example
+     * proj(x as y, b) --> proj(a as x, b, c) =>(a as y, b)
+     * 2. remove used projection in bottom project
+     * @param childProject bottom project
+     * @return project list for merged project
+     */
+    default List<NamedExpression> mergeProjections(Project childProject) {
+        List<NamedExpression> thisProjectExpressions = getProjects();
+        List<NamedExpression> childProjectExpressions = childProject.getProjects();
+        Map<Expression, Expression> bottomAliasMap = childProjectExpressions.stream()
+                .filter(e -> e instanceof Alias)
+                .collect(Collectors.toMap(
+                        NamedExpression::toSlot, e -> e)
+                );
+        return thisProjectExpressions.stream()
+                .map(e -> ExpressionReplacer.INSTANCE.replace(e, bottomAliasMap))
+                .map(NamedExpression.class::cast)
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * replace alias
+     */
+    public static class ExpressionReplacer extends DefaultExpressionRewriter<Map<Expression, Expression>> {
+        public static final ExpressionReplacer INSTANCE = new ExpressionReplacer();
+
+        public Expression replace(Expression expr, Map<Expression, Expression> substitutionMap) {
+            if (expr instanceof SlotReference) {
+                Slot ref = ((SlotReference) expr).withQualifier(Collections.emptyList());
+                return substitutionMap.getOrDefault(ref, expr);
+            }
+            return visit(expr, substitutionMap);
+        }
+
+        /**
+         * case 1:
+         *          project(alias(c) as d, alias(x) as y)
+         *                      |
+         *                      |                          ===>       project(alias(a) as d, alias(b) as y)
+         *                      |
+         *          project(slotRef(a) as c, slotRef(b) as x)
+         * case 2:
+         *         project(slotRef(x.c), slotRef(x.d))
+         *                      |                          ===>       project(slotRef(a) as x.c, slotRef(b) as x.d)
+         *         project(slotRef(a) as c, slotRef(b) as d)
+         * case 3: others
+         */
+        @Override
+        public Expression visit(Expression expr, Map<Expression, Expression> substitutionMap) {
+            if (expr instanceof Alias && expr.child(0) instanceof SlotReference) {
+                // case 1:
+                Expression c = expr.child(0);
+                // Alias doesn't contain qualifier
+                Slot ref = ((SlotReference) c).withQualifier(Collections.emptyList());
+                if (substitutionMap.containsKey(ref)) {
+                    return expr.withChildren(substitutionMap.get(ref).children());
+                }
+            } else if (expr instanceof SlotReference) {
+                // case 2:
+                Slot ref = ((SlotReference) expr).withQualifier(Collections.emptyList());
+                if (substitutionMap.containsKey(ref)) {
+                    Alias res = (Alias) substitutionMap.get(ref);
+                    return res.child();
+                }
+            } else if (substitutionMap.containsKey(expr)) {
+                return substitutionMap.get(expr).child(0);
+            }
+            return super.visit(expr, substitutionMap);
+        }
+    }
 }
+
+
