+++ fe/fe-core/src/test/java/org/apache/doris/nereids/postprocess/MergeProjectPostProcessTest.java
@@ -89,7 +89,6 @@ public class MergeProjectPostProcessTest {
         PhysicalProject proj1 = new PhysicalProject(projList1, placeHolder, proj2);
         MergeProjectPostProcessor processor = new MergeProjectPostProcessor();
         PhysicalPlan newPlan = (PhysicalPlan) proj1.accept(processor, ctx);
-        System.out.println(newPlan.treeString());
         Assertions.assertTrue(newPlan instanceof PhysicalProject);
         Assertions.assertTrue(newPlan.child(0) instanceof PhysicalOlapScan);
         List<NamedExpression> resProjList = ((PhysicalProject<?>) newPlan).getProjects();
